# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Golf Charity Subscription Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithSubscription"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithSubscription"
        "404":
          description: User not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Subscriptions
  /subscriptions:
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/my:
    get:
      operationId: getMySubscription
      tags: [subscriptions]
      summary: Get current user's subscription
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: No subscription
  /subscriptions/{id}/cancel:
    patch:
      operationId: cancelSubscription
      tags: [subscriptions]
      summary: Cancel a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}/renew:
    patch:
      operationId: renewSubscription
      tags: [subscriptions]
      summary: Renew a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # Scores
...
  /winners/{id}/upload-proof:
    post:
      operationId: uploadWinnerProof
      tags: [winners]
      summary: Upload proof image URL for winner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadProofBody"
      responses:
        "200":
          description: Proof uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Winner"
  /winners/{id}/verify:
    patch:
      operationId: verifyWinner
      tags: [winners]
      summary: Approve or reject winner proof (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyWinnerBody"
      responses:
        "200":
          description: Winner verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Winner"
  /winners/{id}/mark-paid:
    patch:
      operationId: markWinnerPaid
      tags: [winners]
      summary: Mark winner as paid (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Winner marked as paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Winner"

  # Admin analytics
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
    UserWithSubscription:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        subscription:
          oneOf:
            - $ref: "#/components/schemas/Subscription"
            - type: "null"
      required:
        - id
        - email
        - name
        - role
        - createdAt
    Subscription:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        plan:
          type: string
        status:
          type: string
        amount:
          type: number
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        renewalDate:
          type: ["string", "null"]
          format: date-time
        stripeSubscriptionId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - plan
        - status
        - amount
        - startDate
        - createdAt
    CreateSubscriptionBody:
      type: object
      properties:
        plan:
          type: string
        paymentMethodId:
          type: ["string", "null"]
      required:
        - plan
...
    Winner:
      type: object
      properties:
        id:
          type: integer
        drawId:
          type: integer
        userId:
          type: integer
        matchTier:
          type: integer
        prizeAmount:
          type: number
        status:
          type: string
        proofUrl:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        userName:
          type: string
        userEmail:
          type: string
        drawMonth:
          type: integer
        drawYear:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - drawId
        - userId
        - matchTier
        - prizeAmount
        - status
        - userName
        - userEmail
        - drawMonth
        - drawYear
        - createdAt
    UploadProofBody:
      type: object
      properties:
        proofUrl:
          type: string
      required:
        - proofUrl
    VerifyWinnerBody:
      type: object
      properties:
        approved:
          type: boolean
        adminNote:
          type: ["string", "null"]
      required:
        - approved
...